ModuleRegistry (#125)
* Move module initialization to registry

* Add imports for Arisa.* configs

* Move registry to main package

* ktlint

* remove unneccasary toList call
